Fix a bug when handling large playability numbers in Limit By Playability condition in CSW12. Also fix a small bug in Limit By Probability.

+++ src/libzyzzyva/SearchForm.cpp
@@ -299,7 +299,7 @@ SearchForm::search()
                 wordItem.setProbabilityOrder(probOrder);
             }
             else if (hasPlayabilityCondition) {
-                int playValue = wordEngine->getPlayabilityValue(
+                qint64 playValue = wordEngine->getPlayabilityValue(
                     lexicon, wordUpper);
                 int playOrder = wordEngine->getPlayabilityOrder(
                     lexicon, wordUpper);
