+++ src/libzyzzyva/WordEngine.h
@@ -70,7 +70,7 @@ class WordEngine : public QObject
         bool isBackHook;
         QString lexiconSymbols;
         QString definition;
-        int playability;
+        qint64 playability;
         ValueOrder playabilityOrder;
         QMap<int, ValueOrder> blankProbabilityOrder;
     };
@@ -85,7 +85,7 @@ class WordEngine : public QObject
         QMap<QString, QMultiMap<QString, QString> > definitions;
         QMap<int, QStringList> stems;
         QMap<QString, int> numAnagramsMap;
-        QMap<QString, int> playabilityMap;
+        QMap<QString, qint64> playabilityMap;
         QMap<int, QSet<QString> > stemAlphagrams;
         mutable QMap<QString, WordInfo> wordCache;
         WordGraph* graph;
@@ -125,7 +125,7 @@ class WordEngine : public QObject
         const;
     QString getBackHookLetters(const QString& lexicon, const QString& word)
         const;
-    int getPlayabilityValue(const QString& lexicon, const QString& word)
+    qint64 getPlayabilityValue(const QString& lexicon, const QString& word)
         const;
     int getPlayabilityOrder(const QString& lexicon, const QString& word)
         const;
